manually declare provider
