+++ example/next.config.js
@@ -1,4 +1,4 @@
-const withImagegen = require('next-plugin-imagegen')
+const { withImagegen } = require('next-plugin-imagegen')
 
 module.exports = withImagegen({
   provider: 'next-plugin-imagegen-puppeteer'
