+++ example/pages/api/imagegen/[...imagegen].js
@@ -1,0 +1,4 @@
+import { handler /* , defaultProvider */ } from 'next-plugin-imagegen'
+import provider from 'next-plugin-imagegen-puppeteer'
+
+export default handler(provider)
