--- example/pages/api/imagegen__.js
@@ -1,1 +1,0 @@
-export {default} from "next-plugin-imagegen/api"
