--- packages/next-plugin-imagegen/.gitignore
@@ -1,2 +1,0 @@
-# generated resources
-api.js
