+++ packages/next-plugin-imagegen/index.js
@@ -1,0 +1,4 @@
+exports.handler = require('./lib/handler')
+exports.provider = require('./lib/default-provider')
+exports.withImagegen = require('./lib/next-config')
+
