+++ packages/next-plugin-imagegen/package.json
@@ -1,10 +1,7 @@
 {
   "name": "next-plugin-imagegen",
   "version": "0.1.1",
-  "main": "./lib/index.js",
-  "scripts": {
-    "prepublish": "rm ./api.js"
-  },
+  "main": "./index.js",
   "files": [
     "lib/*",
     "*.js",
