--- packages/next-plugin-imagegen/provider.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/default-provider')
